Fundamentals of Artificial Intelligence/week5/Project_One/backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn>=0.30,<1.0
transformers>=4.40,<5.0
torch>=2.2
pillow>=10.0
python-multipart>=0.0.9
